+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
-import { Home } from '../Home';
+import { Home } from './Home';
 
 function App() {
   return (
