+++ src/components/App/Home/Body/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Body.js"
+}
