+++ src/components/App/Home/Body/styles.css
@@ -1,0 +1,19 @@
+.ImgHex.shadow {
+  transform-origin: 50% 50%;
+}
+
+.ImgHex {
+  position: absolute;
+  transform-origin: 43% 33%;
+  animation: rotateHexImg 10s linear infinite;
+}
+
+
+@keyframes rotateHexImg {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(359deg);
+  }
+}
