+++ src/components/App/Home/Header/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Header.js"
+}
