+++ src/components/App/Home/Header/styles.css
@@ -1,0 +1,84 @@
+header {
+  position: absolute;
+  z-index: 1;
+  width: 100%;
+  background: indianred;
+}
+
+nav {
+  display: flex;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.15);
+}
+
+h1 {
+  text-transform: uppercase;
+  display: flex;
+  justify-content: center;
+}
+
+em {
+  font: inherit;
+}
+
+ul {
+  list-style: none;
+  display: flex;
+}
+
+ul li {
+  padding: 0 10px;
+}
+
+@keyframes sideSlide {
+  60% {
+    transform: scale(2);
+    color: #808080;
+  }
+
+  80% {
+    transform: scale(2);
+    color: #808080;
+  }
+
+  99% {
+    transform: scale(2.2);
+    color: #FB202D;
+  }
+
+  100% {
+    transform: scale(2);
+    opacity: 1;
+    color: #000000;
+  }
+}
+
+@keyframes showBrand {
+  from {
+    opacity: 0;
+    transform: scale(1);
+  }
+  to {
+    opacity: 1;
+    transform: scale(2);
+  }
+}
+
+.Brand {
+  position: relative;
+  display: inline-block;
+  animation: showBrand 2s forwards;
+}
+
+/* .Brand:hover {
+  animation: sideSlide 2s 2 alternate;
+} */
+
+.BurgerMenu span{
+  width: 35px;
+  height: 5px;
+  background: black;
+  margin: 6px 0;
+  padding: 0;
+  border-radius: 50%;
+  display: block
+}
