+++ src/components/App/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "App.js"
+}
