+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import * as serviceWorker from './serviceWorker';
-import App from './components/App';
+import App from './components/App/App';
 import { GlobalStyle } from './styles/GlobalStyles';
 
 ReactDOM.render(
